Don't reject dash as /bin/sh.
From a discussion with jperkin, who tested this on bulk builds, the
remaining issues with non-BSD echo are in legacy wrappers.

Keep the workaround to use bash on Debian GNU kFreeBSD because it still
uses legacy wrappers and add a comment that that's why it's there.

Now naive usage of bootstrap on many popular Linux distros doesn't require
setting an environment variable.