py-immutabledict: import version 2.0.0
immutable is a fork of frozendict, an immutable wrapper around dictionaries.

It implements the complete mapping interface. It can be used as a drop-in
replacement for dictionaries where immutability is desired. The immutabledict
constructor mimics dict, and all of the expected interfaces (iter, len, repr,
hash, getitem) are provided. Note that an immutabledict does not guarantee the
immutability of its values, so the utility of hash method is restricted by
usage.

The only difference is that the copy() method of immutable takes variable
keyword arguments, which will be present as key/value pairs in the new,
immutable copy.