math/py-scikit-learn: drop ineffective SKLEARN_NO_OPENMP
It is wrong to disable OpenMP and this setting does nothing anyway.
Maybe it did in the past. Now, the build tries to use OpenMP if
possible and this is the right thing to do.

Repeat: This change doesn't affect the build at all since that variable
is not checked (since some time?).