www/drupal8: update to 8.9.16
8.9.16 (2021-05-26)

Maintenance and security release of the Drupal 8 series.

This release fixes a security vulnerability.  Sites are urged to upgrade
immediately after reading the notes below and the security announcement:

* Drupal core - Moderately critical - Cross Site Scripting -
  SA-CORE-2021-003 No other fixes are included.

Versions of Drupal 8 prior to 8.9.x are end-of-life and do not receive
security coverage.


8.9.15 (2021-05-05)

This is a patch (bugfix) release of Drupal 8 and is ready for use on
production sites.  Learn more about Drupal 8.

Drupal 8.9 is the final minor release of the 8.x series.  It is a long-term
support (LTS) version, and will receive security coverage until November
2021.  It provides the same public API as Drupal 9.0 aside from deprecated
code and dependency changes.  (Learn more about Drupal 9.) Note that
features will only be added to Drupal 9 minor releases, so plan to adopt
Drupal 9 as soon as possible so that you can easily update to Drupal 9.2 and
later.

If you are upgrading to this release from 8.8.x, read the Drupal 8.9.0
release notes before you upgrade.

Known issues

Search the issue queue for known issues.

Important changes

The default glossary view did not previously include a filter to exclude
unpublished content.  This view now includes such a filter by default, and
an update function is provided with this release to add a status filter to
the view on existing installations which do not have it.

Dependency updates

The composer/composer development dependency has been updated from 1.10.6 to
1.10.22.

Archive_Tar has been updated to 1.4.13 for security hardening.

Drupal core's development dependency on the Nightwatch npm package has been
increased from 1.2.1 to 1.6.3 and all locked versions of dependencies have
been updated to address security issues in these dependencies.
The minimum version of node.js for 8.9.x development has been increased to
version 10.

Underscore.js has been updated to 1.13.1